requirements.txt
streamlit==0.61.0
numpy==1.18.5
tensorflow==2.2.0
pandas==0.24.2
Pillow==7.2.0
